Import initial master and slave version
We begin with a master that will handle requests with goroutines and
thus can handle multiple slaves at once.
It can be gracefully stopped with SIGINT or SIGTERM, it will wait
all the connections to complete before dropping them.

The slave registers itself as soon as it starts, failing to do so
will exit the program; no retry for now.
Handles SIGINT and SIGTERM nicely.

The master keeps track of registered slaves and warn about double
registrations.

Colored logs on standard output and a file named after the
executable.