Switch to a proper protocol
By design we want a full duplex protocol, with heartbeat and TLS.
Instead of using our homeground code or resort to net/rpc to encode and
decode messages we switch to gRPC (http://www.grpc.io/): a high performance
RPC framework on HTTP/2.

gRPC clients knows when the connection goes down, this allows slaves to
reconnect when the master is restarted - which means less maintainance
burden compared to Builder 1.0.

While we are here we relocate code to the root directory and refactor so
we now have master and slave in separate packages and a single executable
that can launch either of them.

This simplifies deployment which is now a matter of copying an executable
and configure it.

Let slave name be overridden
Also, users can override the slave name with a command line arguments,
making testing more agile.