Send package information to slave when scheduling a build
Users now must specify the architecture when scheduling a build with
builder-cli.

Master now retrieves package information from the database and sends it
to the slave when dispatching the job.

Jobs are still dispatched to the first slave available regardless the
channel and architectures supported.

In a followup commit we'll be supporting building images too.