Web server
Web server now serves web socket connections as well as static content
and template pages.

A template page initiate the communication with the web socket and
receives messages as events are generated (such as new jobs queued,
completed or failed).

The page can ask specific information, for example subscribed slaves,
package information. It can also send actions such as build requests.

Server-side we have a hub that maintains the set of active connections
and broadcasts messages to the connections.