Dispatch jobs according to topics
A slave subscribe to the master for some topics: job type and
architectures.

These topics determine what the slave is capable to do.
For example a x86 slave will never be able to build ARM packages
or images, hence it will only subscribe to i386 and/or x86_64
topics.

Topics were previously called channels, they are renamed to topics
to avoid confusion with Go channels.

Jobs must be dispatched according to the topic, otherwise the build
will almost certainly fail.