Refactor the protocol
Subscribe now returns immediately, this way slave can create a valid
context and make it available to every function right away.

PickJob is called by slave to initiate a communication.
Master streams job requests to slave which in turn streams updates
back to master.

Upload has also been changed a little bit, that is chunk hash is no
longer sent because we only evaulate that at the end of the
transmission.